requirements.txt
appdirs==1.4.4
astroid==2.4.1
attrs==19.3.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
Cython==0.29.17
decorator==4.4.2
future==0.18.2
idna==2.10
imagecodecs==2020.2.18
imageio==2.8.0
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
networkx==2.4
numpy==1.18.4
pathspec==0.8.0
Pillow==7.1.2
pycocotools==2.0
pylint==2.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
regex==2020.5.7
requests==2.24.0
rope==0.17.0
scikit-image==0.17.2
scipy==1.4.1
six==1.14.0
tifffile==2020.5.11
toml==0.10.0
torch==1.5.0
torchvision==0.6.0
typed-ast==1.4.1
urllib3==1.25.9
wrapt==1.12.1
